pyproject.toml
[project]
name = "mcp-server-with-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "fastmcp>=2.10.3",
    "httpx>=0.28.1",
    "mcp[cli]>=1.10.1",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "starlette>=0.46.2",
    "uvicorn>=0.35.0",
]
